web-ui/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-groq>=1.0.0
langchain-community>=0.0.20
langgraph>=0.2.0
